refactor: move http types mod
