+++ src/satellite/src/storage/certification/cert.rs
@@ -3,7 +3,7 @@ use crate::storage::certification::constants::{
 };
 use crate::storage::certification::tree_utils::response_headers_expression;
 use crate::storage::certification::types::certified::CertifiedAssetHashes;
-use crate::storage::types::http::HeaderField;
+use crate::storage::http::types::HeaderField;
 use crate::types::core::Blob;
 use base64::encode;
 use ic_cdk::api::{data_certificate, set_certified_data};
