+++ src/satellite/src/storage/certification/impls.rs
@@ -11,7 +11,7 @@ use crate::storage::constants::{
     ROOT_404_HTML, ROOT_INDEX_HTML, ROOT_PATH,
 };
 use crate::storage::http::headers::{build_asset_headers, build_redirect_headers};
-use crate::storage::types::http::HeaderField;
+use crate::storage::http::types::HeaderField;
 use crate::storage::types::state::FullPath;
 use crate::storage::types::store::Asset;
 use crate::storage::url::alternative_paths;
