+++ src/satellite/src/storage/http/mod.rs
@@ -1,3 +1,4 @@
 pub mod headers;
 pub mod response;
+pub mod types;
 pub mod utils;
